+++ test/unit/cli.test.ts
@@ -4,7 +4,8 @@ import * as net from "net"
 import * as os from "os"
 import * as path from "path"
 import { Args, parse, setDefaults, shouldOpenInExistingInstance } from "../../src/node/cli"
-import { paths, tmpdir } from "../../src/node/util"
+import { tmpdir } from "../../src/node/constants"
+import { paths } from "../../src/node/util"
 
 type Mutable<T> = {
   -readonly [P in keyof T]: T[P]
