+++ test/unit/socket.test.ts
@@ -4,8 +4,9 @@ import * as net from "net"
 import * as path from "path"
 import * as tls from "tls"
 import { Emitter } from "../../src/common/emitter"
+import { tmpdir } from "../../src/node/constants"
 import { SocketProxyProvider } from "../../src/node/socket"
-import { generateCertificate, tmpdir } from "../../src/node/util"
+import { generateCertificate } from "../../src/node/util"
 
 describe("SocketProxyProvider", () => {
   const provider = new SocketProxyProvider()
